requirements.txt
blinker >= 1.4
nose >= 1.3.7
pillow >= 3.2.0
requests >= 2.10.0
tensorflow >= 2.0.0